requirements.txt
python-fasthtml
uvicorn[standard]